Add humanTactics static class
