+++ UriSudokuSolver/SolvingAlgorithm/SudokuSolverUtility.cs
@@ -106,7 +106,7 @@ namespace UriSudokuSolver
 
 
         /*Get valid values for a given cell.*/
-        private static int GetValidValuesForCell(byte[,] board, int[] validValuesRow, int[] validValuesColumn, int[] validValuesBox, int sqrSize, int row, int col)
+        public static int GetValidValuesForCell(byte[,] board, int[] validValuesRow, int[] validValuesColumn, int[] validValuesBox, int sqrSize, int row, int col)
         {
             // OR bit operator on the unvalid values (bit in the index - 1 of the value is 1) of the row, column and box representing the valid values of the cell
             int nonValidValuesRowColSqr = validValuesRow[row] | validValuesColumn[col] | validValuesBox[(row / sqrSize) * sqrSize + col / sqrSize];
@@ -128,7 +128,7 @@ namespace UriSudokuSolver
 
 
         /*Count the number of bits in a binary number.*/
-        private static int CountBits(int value)
+        public static int CountBits(int value)
         {
             int count = 0;
             // check how many bits by AND operation with the value and the value - 1 (example 00000100 & 00000011 = 00000000)
@@ -177,63 +177,9 @@ namespace UriSudokuSolver
             validValuesBox[(row / sqrSize) * sqrSize + col / sqrSize] &= ~masks[value - 1];
         }
 
-        /*Try to solve board once with human tactics*/
-        public static int HumanTactics(Stack<int> savedValues, byte[,] board, int[] validValuesRow, int[] validValuesColumn, int[] validValuesBox, int[] masks, int sqrSize)
-        {
-            // how many changes made
-            int found = 0;
-            int validValues, bitCount, hiddenSingle;
-            // Go over all the matrix and search empry spaces
-            for (int row = 0; row < board.GetLength(0); row++)
-            {
-                for (int col = 0; col < board.GetLength(1); col++)
-                {
-                    // if empty space
-                    if (board[row, col] == 0)
-                    {
-                        //get the values that this cell can get (in binary)
-                        validValues = GetValidValuesForCell(board, validValuesRow, validValuesColumn, validValuesBox, sqrSize, row, col);
-                        // count them
-                        bitCount = CountBits(validValues);
-                        if (bitCount == 1)
-                        {
-                            UpdateAndPush(savedValues, board, validValuesRow, validValuesColumn, validValuesBox, masks, sqrSize, row, col, validValues);
-                            found++;
-                            continue;
-                        }
-                        if (bitCount == 0)
-                        {
-                            // if is empty and doesn't have valid values then the board is unsolvale
-                            CleanStack(board, savedValues, validValuesRow, validValuesColumn, validValuesBox, masks, sqrSize, found);
-                            return -1;
-                        }
-                        // try to search for hidden singles --> for cells that have an option that no other cell in the row, column or box can have
-                        hiddenSingle = HiddenSingles(board, validValuesRow, validValuesColumn, validValuesBox, masks, sqrSize, row, col, validValues);
-                        if (hiddenSingle > 0)
-                        {
-                            UpdateAndPush(savedValues, board, validValuesRow, validValuesColumn, validValuesBox, masks, sqrSize, row, col, hiddenSingle);
-                            found++;
-                        }
-                        if (hiddenSingle == -1)
-                        {
-                            // if is empty and can't get valid values then the board is unsolvale
-                            CleanStack(board, savedValues, validValuesRow, validValuesColumn, validValuesBox, masks, sqrSize, found);
-                            return -1;
-                        }
-                    }
-                }
-            }
-            return found;
-        }
 
-        /*Update the board with a value and save the changes in the stack*/
-        private static void UpdateAndPush(Stack<int> savedValues, byte[,] board, int[] validValuesRow, int[] validValuesColumn, int[] validValuesBox, int[] masks, int sqrSize, int row, int col, int value)
-        {
-            UpdateBoard(board, validValuesRow, validValuesColumn, validValuesBox, masks, sqrSize, row, col, value);
-            savedValues.Push(row * board.GetLength(0) + col);
-        }
 
-        /*Cleans the stack*/
+        /*Cleans the stack form saved values and return them to the board*/
         public static void CleanStack(byte[,] board, Stack<int> validStack, int[] validValuesRow, int[] validValuesColumn, int[] validValuesBox, int[] masks, int sqrSize, int cellNum)
         {
             int index, row, col;
@@ -243,57 +189,11 @@ namespace UriSudokuSolver
                 row = index / board.GetLength(0);
                 col = index % board.GetLength(0);
                 //deletes from valide values arrays
-                AddValueToValidValues(board, validValuesRow, validValuesColumn, validValuesBox, masks, sqrSize, row, col);
+                SudokuSolverUtility.AddValueToValidValues(board, validValuesRow, validValuesColumn, validValuesBox, masks, sqrSize, row, col);
                 board[row, col] = 0;
             }
 
         }
-        /*The function get board, row and column of an empty cell and seek for its only value as naked single --> if it can't be in any other row col and box it has to be there.*/
-        public static int HiddenSingles(byte[,] board, int[] validValuesRow, int[] validValuesColumn, int[] validValuesBox, int[] masks, int sqrSize, int row, int col, int validValues)
-        {
-            // Get the cell index in the box.
-            int cellInBox = row % sqrSize * sqrSize + col % sqrSize;
-            int possibleInOtherCells = 0;
-            int rowInsideBox, colInsideBox, possibleForCell;
-            // check the box number
-            int boxNumber = (row - row % sqrSize) + (col - col % sqrSize) / sqrSize;
-            //Go over all the boxes in the board
-            for (int i = 0; i < board.GetLength(0); i++)
-            {
-                //if the box is not the current box
-                if (i != cellInBox)
-                {
-                    //get the row and column of the cell inside the box
-                    rowInsideBox = sqrSize * (boxNumber / sqrSize) + i / sqrSize;
-                    colInsideBox = (boxNumber % sqrSize) * sqrSize + i % sqrSize;
-                    //if the cell is empty
-                    if (board[rowInsideBox, colInsideBox] == 0)
-                    {
-                        // add the valid values of the cell to the possible values for all the cells in the box
-                        possibleInOtherCells |= GetValidValuesForCell(board, validValuesRow, validValuesColumn, validValuesBox, sqrSize, rowInsideBox, colInsideBox);
-                    }
-                }
-
-            }
-            // get the possible values for the cell by using the NOT and OR bit operators to remove the possible values for all the cells in the box
-            possibleForCell = ~(~validValues | possibleInOtherCells);
-
-            //if the cell has only one valid value
-            if (CountBits(possibleForCell) == 1)
-            {
-                
-                return possibleForCell;
-            }
-            
-            // if the cell has no valid values
-            if (possibleForCell == 0)
-                return 0;
-            return -1;
-
-
-        }
-
-
